Refactor everything to Trunk
