+++ gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.jvmargs=-Xmx2G
 
-project.name=SpruceLaunchwrapper
+project.name=Trunk
 project.version=1.2.1
 project.group=xyz.spruceloader
