+++ src/main/java/xyz/spruceloader/trunk/api/ArgumentMap.java
@@ -1,4 +1,4 @@
-package xyz.spruceloader.launchwrapper.api;
+package xyz.spruceloader.trunk.api;
 
 import java.util.*;
