+++ src/main/java/xyz/spruceloader/trunk/api/LaunchTransformer.java
@@ -1,13 +1,13 @@
-package xyz.spruceloader.launchwrapper.api;
+package xyz.spruceloader.trunk.api;
 
-import xyz.spruceloader.launchwrapper.LaunchClassLoader;
+import xyz.spruceloader.trunk.TrunkClassLoader;
 
 public interface LaunchTransformer {
 
     default void takeArguments(ArgumentMap argMap, EnvSide env) {
     }
 
-    default void injectIntoClassLoader(LaunchClassLoader classLoader) {
+    default void injectIntoClassLoader(TrunkClassLoader classLoader) {
     }
 
     default byte[] transform(String className, byte[] rawClass) {
