+++ src/main/java/xyz/spruceloader/trunk/exceptions/LoadingException.java
@@ -1,4 +1,4 @@
-package xyz.spruceloader.launchwrapper.exceptions;
+package xyz.spruceloader.trunk.exceptions;
 
 public class LoadingException extends RuntimeException {
     public LoadingException(String message, Throwable cause) {
